Grid and Linear menus are now under a single implementation.
Some functions are simply guarded into two if statements based on whether or not they're gridmenu or not. A later implementation will include (if i choose to do so) the ability to override the default grid behaviour for specific elements, i.e. pressing left from a certain element will always lead to another specific element. This might be a lot more complicated than I thought so I might not bother.