Added Sprite and SpriteAnimation
Now users can define a Sprite as an atlas and a corresponding Source Rectangle instead of needing to maintain Rectangles and Texture2Ds independently. Originally Sprite used a Texture2D reference to save space but by the time the app context shifts to DrawSpriteAnimation the dereferenced Texture2D was unrecognizable by the GPU and nothing would draw. Ideally we would find a cheaper way to store Texture2D references.